zero-shot-prompting/requirements.txt
numpy
pandas
PyYAML
tqdm
Pillow
requests
datasets
decord
opencv-python
loguru
nltk
